requirements.txt
Django==4.1.3
djangorestframework==3.14.0
python-environ==0.4.54
drf-spectacular==0.24.2
djangorestframework-simplejwt==5.2.2
pyjwt==2.6.0
cryptography==38.0.4
psycopg2==2.9.5
django-debug-toolbar==3.8.1